Updating images to call artworks

+++ git-artsy-backend/src/main/java/org/launchcode/git_artsy_backend/repositories/ImageSearchRepository.java
@@ -4,6 +4,6 @@ import org.launchcode.git_artsy_backend.models.ImageSearch;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ImageSearchRepository extends JpaRepository<ImageSearch, Long> {
+public interface ImageSearchRepository extends JpaRepository<ImageSearch, Integer> {
     List<ImageSearch> findByTitleContainingOrDescriptionContaining(String title, String description);
 }
